Made major performance improvements, cleanup, refactoring, and added better dependency injection
