+++ core/ClassInspector.py
@@ -1,7 +1,6 @@
 import enum
 import importlib
 import inspect
-import json
 import pkgutil
 
 from langchain import OpenAI
@@ -25,240 +24,7 @@ def get_class_attributes(cls):
     return attributes
 
 
-def get_langchain_example_from_llm():
-    prompt_template = f"""You\'re a helpful assistant. I\'ll give you example code that uses some Python libraries I\'m interested in, and I want you to provide me with a code example of RAG on langchain with the OpenAI LLM and AstraDB as my vector store. Don\'t give me any explanation or description. Just give me the code.
-
-EXAMPLES:
-
-
-EXAMPLE NOTEBOOK 1:
-
-RAG
-Let’s look at adding in a retrieval step to a prompt and LLM, which adds up to a “retrieval-augmented generation” chain
-
-!pip install langchain openai faiss-cpu tiktoken
-
-from operator import itemgetter
-
-from langchain.chat_models import ChatOpenAI
-from langchain.embeddings import OpenAIEmbeddings
-from langchain.prompts import ChatPromptTemplate
-from langchain.vectorstores import FAISS
-from langchain_core.output_parsers import StrOutputParser
-from langchain_core.runnables import RunnableLambda, RunnablePassthrough
-
-vectorstore = FAISS.from_texts(
-    [\"harrison worked at kensho\"], embedding=OpenAIEmbeddings()
-)
-retriever = vectorstore.as_retriever()
-
-template = \"\"\"Answer the question based only on the following context:
-{{context}}
-
-Question: {{question}}
-\"\"\"
-prompt = ChatPromptTemplate.from_template(template)
-
-model = ChatOpenAI()
-
-chain = (
-    {{\"context\": retriever, \"question\": RunnablePassthrough()}}
-    | prompt
-    | model
-    | StrOutputParser()
-)
-
-chain.invoke(\"where did harrison work?\")
-
-\'Harrison worked at Kensho.\'
-
-template = \"\"\"Answer the question based only on the following context:
-{{context}}
-
-Question: {{question}}
-
-Answer in the following language: {{language}}
-\"\"\"
-prompt = ChatPromptTemplate.from_template(template)
-
-chain = (
-    {{
-        \"context\": itemgetter(\"question\") | retriever,
-        \"question\": itemgetter(\"question\"),
-        \"language\": itemgetter(\"language\"),
-    }}
-    | prompt
-    | model
-    | StrOutputParser()
-)
-
-chain.invoke({{\"question\": \"where did harrison work\", \"language\": \"italian\"}})
-
-\'Harrison ha lavorato a Kensho.\'
-
-
-EXAMPLE NOTEBOOK 2:
-
-# Get Wikipedia data
-
- !wget https://raw.githubusercontent.com/GeorgeCrossIV/Langchain-Retrieval-Augmentation-with-CASSIO/main/20220301.simple.csv
-
-Import the 20220301.simple wikipedia from the CSV file
-
-import pandas as pd
-
-data = pd.read_csv(\'20220301.simple.csv\')
-
-There are 10,000 entries in the Wikipedia data file. We\'ll reduce the dataset to 10 rows for this demo. It takes a while to process the data; however, feel free to increase the number of rows for future demo runs.
-
-data = data.head(10)
-data = data.rename(columns={{\'text \': \'text\'}})
-data
-
-We will execute queries against the [Andouille](https://simple.wikipedia.org/wiki/Andouille) Wikipedia entries later in this demo. The Wikipedia data used in this demo is from a snapshot in time, stored in a CSV file. Below is the text of the Wikipedia record that will be processed.
-
-data.iloc[9][\'text\']
-
-## Compute vector embedding using MiniLM
-
-from langchain.embeddings import HuggingFaceEmbeddings
-from langchain.vectorstores import AstraDB as AstraDB
-embeddings = HuggingFaceEmbeddings(model_name=\"all-MiniLM-L12-v2\")
-
-### Setup AstraDB vector store instance for LangChain
-
-cassandraVectorStore = AstraDB(
-    embedding=embeddings,
-    collection_name=ASTRA_COLLECTION,
-    token=ASTRA_DB_TOKEN,
-    api_endpoint=ASTRA_DB_API_ENDPOINT,
-)
-
-# Build table with embeddings
-
-from langchain.docstore.document import Document
-from langchain.text_splitter import RecursiveCharacterTextSplitter
-
-
-mySplitter = RecursiveCharacterTextSplitter(chunk_size=250, chunk_overlap=120)
-
-def create_vector_index(row, myCassandraVStore):
-  metadata = {{
-    \'url\': row[\'url\'],
-    \'title\': row[\'title\']
-  }}
-  page_content = row[\'text\']
-
-  wikiDocument = Document(
-      page_content=page_content,
-      metadata=metadata
-  )
-  wikiDocs = mySplitter.transform_documents([wikiDocument])
-  myCassandraVStore.add_documents(wikiDocs)
-
-for index, row in data.iterrows():
-  create_vector_index(row, cassandraVectorStore)
-
-from ibm_watson_machine_learning.foundation_models import Model
-from ibm_watson_machine_learning.foundation_models.extensions.langchain import WatsonxLLM
-
-model_id = \"google/flan-t5-xxl\"
-
-credentials = {{
-    \"url\"    : f\"https://{{IBM_REGION}}.ml.cloud.ibm.com\",
-    \"apikey\" : apiSecret
-}}
-
-gen_parms = {{
-    \"DECODING_METHOD\" : \"greedy\",
-    \"MIN_NEW_TOKENS\" : 1,
-    \"MAX_NEW_TOKENS\" : 50
-}}
-
-model = Model( model_id, credentials, gen_parms, PROJECT_ID )
-lang_chain_compatible_model = WatsonxLLM(model=model)
-
-
-results = cassandraVectorStore.similarity_search(\"What is Andouille?\", 1)
-
-def print_results(exampleRows):
-  for row in exampleRows:
-    print(f\"\"\"{{row.page_content}}\n\"\"\")
-
-print_results(results)
-
-results = cassandraVectorStore.similarity_search(\"What temperature should Andouille be cooked?\", 3)
-print_results(results)
-
-We now have the relevant context that we can leverage to answer our question, but it\'s not super readable. Let\'s use IBM\'s watsonx chat completion API to provide a more readable answer.
-
-First, let\'s prep the prompt.
-
-The prompt will be prepared like this behind the scenes by LangChain:
-
-```
-def prepare_prompt(context, question):
-
-  context_as_string =\'\n\'.join([row.body_blob for row in context])
-  prompt_template = f\"\"\"Use the following pieces of context to answer the question at the end. If you don\'t know the answer, just say that you don\'t know, don\'t try to make up an answer.
-
-  {{context_as_string}}
-
-  Question: {{question}}
-  Helpful Answer:\"\"\"
-  return prompt_template
-
-prompt = prepare_prompt(results, question)
-```
-(This prompt template was derived from [this LangChain module](https://github.com/devinbost/langchain/blob/f35db9f43e2301d63c643e01251739e7dcfb6b3b/libs/langchain/langchain/chains/question_answering/stuff_prompt.py#L10).)
-
-from langchain.chains import RetrievalQA
-
-chain = RetrievalQA.from_chain_type(
-        llm=lang_chain_compatible_model,
-        chain_type=\"stuff\",
-        retriever=cassandraVectorStore.as_retriever(),
-        return_source_documents=True
-)
-
-### Create helper function to format the output:
-
-def format_output(lc_result):
-  print(f\"QUERY: {{lc_result[\'query\']}}\")
-  print(f\"RESPONSE: {{lc_result[\'result\']}}\")
-  print(f\"\n\nSOURCE DOCUMENTS:\n\")
-  for doc in lc_result[\"source_documents\"]:
-    print(f\"TITLE: {{doc.metadata[\'title\']}}\")
-    print(f\"URL: {{doc.metadata[\'url\']}}\")
-
-### Run watsonx via LangChain on a given question
-
-Now that we have a prepared prompt and the relevant context, let\'s ask watsonx to give us a more readable answer.
-
-output = chain({{\"query\":\"What temperature should Andouile be cooked?\"}})
-format_output(output)
-
-Notice that if we were asking a more difficult question across more source documents, the RAG integration would pull insights from muliple sources to give a more complete summary answer.
-
-Regardless, let\'s compare our results to the results of running the model without the context provided by vector search and RAG.
-
-import json
-
-generated_response = model.generate(\"What temperature should Andouile be cooked?\", None )
-print( json.dumps( generated_response[\"results\"][0][\"generated_text\"], indent=2 ) )
-
-The \"\u00b0\" is unicode for the ° character, so the model without RAG gave us 190°C, which is 374 degrees Fahrenheit - quite different from the temperature provided by retrieval augmented generation!
-
-374 degrees Fahrenheit might make sense for a more typical meat cooking temperature, but the critical differentiator here is that Andouille sausage is a smoked sausage by definition, which is clear from the context we retrieve via vector search. So, the correct answer was actually obtained when we leveraged vector search to provide the context necessary for watsonx to give us the desired answer.
-
-### RAG is a powerful tool
-
-To get more information, please reach out to us at Datastax or contact your IBM rep for more information about how we can partner with you to leverage RAG for your Generative AI use case.
-[Check out more at our partnership page.](https://www.datastax.com/partners/ibm)"""
-    chain = build_prompt_from_template(prompt_template)
-    result = chain.invoke({})
-
-
+@DeprecationWarning
 def get_class_attributes_from_llm(cls):
     prompt_template = """You're a helpful assistant. When given the following Python source code, give me the attributes 
 mentioned in the provided code and comments (if any) as a JSON object with the name, type, and default (if any) of each
