+++ core/Config.py
@@ -4,10 +4,10 @@ import logging
 
 class Config:
     def __init__(self):
-        self.scratch_path = os.getenv("SCRATCH_PATH", "scratch")
-        self.data_path = os.getenv("DATA_PATH", "data")
+        self.scratch_path = os.getenv("SCRATCH_PATH", "../scratch")
+        self.data_path = os.getenv("DATA_PATH", "../data")
         self.openai_json = os.getenv(
-            "OPENAI_JSON", "scratch/devin.bost@datastax.com-token.json"
+            "OPENAI_JSON", "../scratch/devin.bost@datastax.com-token.json"
         )
         self.keyspace_name = os.getenv("KEYSPACE_NAME", "keyspace")
         self.table_name = os.getenv("TABLE_NAME", "table")
@@ -17,14 +17,6 @@ class Config:
     def create_directories(self):
         os.makedirs(self.scratch_path, exist_ok=True)
         os.makedirs(self.data_path, exist_ok=True)
-
-    def secure_connect_bundle_path(self):
-        return os.path.join(
-            self.scratch_path, f"secure-connect-{self.database_name}.zip"
-        )
-
-    def get_secure_bundle_full_path(self):
-        return os.path.join(os.getcwd(), self.secure_connect_bundle_path())
 
 
 config = Config()
