+++ core/Crawler.py
@@ -1,9 +1,8 @@
 import asyncio
-import time
-from typing import Tuple, List
 from aiohttp import ClientSession
 from bs4 import BeautifulSoup
 import requests
+from langchain_community.vectorstores.astradb import AstraDB
 from newspaper import Article
 from langchain.docstore.document import Document
 import aiohttp
@@ -12,8 +11,6 @@ from DataAccess import DataAccess
 import logging
 from datetime import datetime
 
-from Config import config
-from pydantic import BaseModel
 from typing import List
 
 from pydantic_models.PageContent import PageContent
@@ -134,7 +131,7 @@ class Crawler:
         url: str,
         progress_bar: DeltaGenerator,
         session: ClientSession,
-        table_name: str,
+        vector_store: AstraDB,
     ):
         """
         Asynchronously handles processing of a single URL.
@@ -142,7 +139,7 @@ class Crawler:
             url (str): The URL to process.
             progress_bar (DeltaGenerator): Streamlit UI element for progress indication.
             session (ClientSession): The HTTP client session for making requests.
-            table_name (str): The name of the table where data is to be stored.
+            vector_store (AstraDB): The LangChain AstraDB instance where data is to be stored.
         """
         async with self.semaphore:  # this will wait if there are already too many tasks running:
             page_content = await self.async_chunk_page(url, session)
@@ -180,9 +177,6 @@ class Crawler:
                     split_docs = self.data_access.splitter.transform_documents(
                         page_docs
                     )
-                    vector_store = self.data_access.getVectorStore(
-                        table_name=table_name
-                    )
                     # vector_store.aadd_documents(split_docs) isn't yet implemented. We will work around it.
                     # await vector_store.aadd_documents(split_docs)
                     loop = asyncio.get_event_loop()
@@ -198,9 +192,9 @@ class Crawler:
             # Only call the UI update if no update is currently in progress to avoid blocking threads with UI updates since it's okay if some get skipped:
             if not self.ui_update_in_progress:
                 self.ui_update_in_progress = True
-            await self.update_UI(self.counter, progress_bar)
+            await self.update_ui(self.counter, progress_bar)
 
-    async def update_UI(self, counter: int, progress_bar: DeltaGenerator):
+    async def update_ui(self, counter: int, progress_bar: DeltaGenerator):
         """
         Asynchronously updates the user interface with the progress of crawling.
         Parameters:
@@ -220,12 +214,12 @@ class Crawler:
         )
         self.ui_update_in_progress = False
 
-    async def process_urls(self, progress_bar: DeltaGenerator, table_name: str):
+    async def process_urls(self, progress_bar: DeltaGenerator, vector_store: AstraDB):
         """
         Asynchronously processes a list of URLs.
         Parameters:
             progress_bar (DeltaGenerator): Streamlit UI element for progress indication.
-            table_name (str): The name of the table where data is to be stored.
+            vector_store (AstraDB): The instance of LangChain AstraDB used to store the crawled data.
         """
 
         timeout = aiohttp.ClientTimeout(
@@ -235,62 +229,93 @@ class Crawler:
             timeout=timeout
         ) as session:  # If needed, use session for HTTP requests
             tasks = [
-                self.handle_url(url, progress_bar, session, table_name)
+                self.handle_url(url, progress_bar, session, vector_store)
                 for url in self.urls
             ]
             await asyncio.gather(*tasks)
 
+    def _initialize_crawl(
+        self, sitemap_urls: List[str] | str, is_list: bool = False
+    ) -> None:
+        """
+        Initializes the crawling process by extracting and setting unique URLs.
+
+        Parameters:
+            sitemap_urls: A string or a list of strings representing the sitemap URLs to crawl.
+            is_list: A boolean flag to indicate if sitemap_urls is a list of URLs.
+
+        Returns:
+            None: This method doesn't return anything but sets self.urls to the unique URLs.
+        """
+        if self.urls is None:
+            if is_list:
+                all_urls = [
+                    url
+                    for sitemap in sitemap_urls
+                    for url in self.get_sitemap_urls(sitemap, onlyEnglish=True)
+                    if "espanol" not in url
+                ]
+                self.urls = list(set(all_urls))
+            else:
+                self.urls = self.get_sitemap_urls(sitemap_urls, onlyEnglish=True)
+
+            length_of_urls = len(self.urls)
+            logging.info(f"Length of unique URLs is: {length_of_urls}")
+            print(f"Length of unique URLs is: {length_of_urls}")
+
+    def _execute_crawl(
+        self, progress_bar: DeltaGenerator, vector_store: AstraDB
+    ) -> None:
+        """
+        Executes the crawling and ingestion process in an asynchronous loop.
+
+        Parameters:
+            progress_bar: A Streamlit UI element for progress indication.
+            vector_store: An instance of AstraDB to store the crawled data.
+
+        Returns:
+            None: This method sets up and runs the asynchronous event loop.
+        """
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        try:
+            loop.run_until_complete(self.process_urls(progress_bar, vector_store))
+        finally:
+            loop.close()
+
     def async_crawl_and_ingest(
-        self, sitemap_url: str, progress_bar: DeltaGenerator, table_name: str
-    ):
+        self, sitemap_url: str, progress_bar: DeltaGenerator, vector_store: AstraDB
+    ) -> None:
         """
         Asynchronously crawls and ingests data from a given sitemap URL.
-        Parameters:
-            sitemap_url (str): The sitemap URL to crawl.
-            progress_bar (DeltaGenerator): Streamlit UI element for progress indication.
-            table_name (str): The name of the table where data is to be stored.
-        """
-        if self.urls is None:
-            # TODO: Need to cache the following step:
-            self.urls = self.get_sitemap_urls(sitemap_url, onlyEnglish=True)
 
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-        try:
-            loop.run_until_complete(self.process_urls(progress_bar, table_name))
-        finally:
-            loop.close()
+        Parameters:
+            sitemap_url: The sitemap URL to crawl.
+            progress_bar: Streamlit UI element for progress indication.
+            vector_store: An instance of AstraDB to store the crawled data.
+
+        Returns:
+            None: This method orchestrates the crawling and ingestion process.
+        """
+        self._initialize_crawl([sitemap_url])
+        self._execute_crawl(progress_bar, vector_store)
 
     def async_crawl_and_ingest_list(
-        self, sitemap_url_list: List[str], progress_bar: DeltaGenerator, table_name: str
-    ):
+        self,
+        sitemap_url_list: List[str],
+        progress_bar: DeltaGenerator,
+        vector_store: AstraDB,
+    ) -> None:
         """
         Asynchronously crawls and ingests data from a list of sitemap URLs.
-        Parameters:
-            sitemap_url_list (List[str]): A list of sitemap URLs to crawl.
-            progress_bar (DeltaGenerator): Streamlit UI element for progress indication.
-            table_name (str): The name of the table where data is to be stored.
-        """
-        if self.urls is None:
-            # TODO: Need to cache the following step:
-            all_urls = [
-                url
-                for sitemap in sitemap_url_list
-                for url in self.get_sitemap_urls(sitemap, onlyEnglish=True)
-                if "espanol" not in url
-            ]
-            length_of_urls = len(all_urls)
-            unique_urls = list(set(all_urls))
-            logging.info(f"Length of all_urls is: {length_of_urls}")
-            print(f"Length of all_urls is: {length_of_urls}")
-            length_of_unique_urls = len(unique_urls)
-            print(f"Length of unique all_urls is: {length_of_unique_urls}")
-            logging.info(f"Length of unique all_urls is: {length_of_unique_urls}")
-            self.urls = unique_urls
 
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-        try:
-            loop.run_until_complete(self.process_urls(progress_bar, table_name))
-        finally:
-            loop.close()
+        Parameters:
+            sitemap_url_list: A list of sitemap URLs to crawl.
+            progress_bar: Streamlit UI element for progress indication.
+            vector_store: An instance of AstraDB to store the crawled data.
+
+        Returns:
+            None: This method orchestrates the crawling and ingestion process for multiple URLs.
+        """
+        self._initialize_crawl(sitemap_url_list, is_list=True)
+        self._execute_crawl(progress_bar, vector_store)
