--- requirements.txt
@@ -1,20 +1,0 @@
-aiohttp~=3.8.5
-bs4~=4.12.2
-cassio
-ibm-watson-machine-learning
-newspaper3k
-pandas~=1.5.3
-requests~=2.31.0
-sentence-transformers
-streamlit~=1.29.0
-typing~=3.10.0.0
-wget~=3.2
-ragstack-ai
-beautifulsoup4~=4.12.2
-fastapi~=0.105.0
-cassandra-driver>=3.28.0
-astrapy>=0.6.2
-graphviz~=2.50.0
-pydantic~=2.3.0
-sortedcontainers~=2.4.0
-nltk
