+++ tests/test_chains.py
@@ -6,21 +6,18 @@ from langchain.chat_models import ChatOpenAI
 from langchain.embeddings import OpenAIEmbeddings
 from langchain.prompts import ChatPromptTemplate
 from langchain.vectorstores import FAISS
-from langchain_core.output_parsers import StrOutputParser, ListOutputParser
+from langchain_core.output_parsers import StrOutputParser
 from langchain_core.runnables import (
-    RunnablePassthrough,
     RunnableLambda,
     RunnableParallel,
     Runnable,
 )
 
 import unittest
-from unittest.mock import MagicMock, patch
 
-from ChainFactory import ChainFactory
-import PromptFactory
-from DataAccess import DataAccess
-import uuid
+from core.ChainFactory import ChainFactory
+from core import PromptFactory
+from core.DataAccess import DataAccess
 
 from pydantic_models.ColumnSchema import ColumnSchema
 from pydantic_models.PropertyInfo import PropertyInfo
@@ -136,7 +133,7 @@ class TestChains(unittest.TestCase):
                 ColumnSchema(column_name="transcript", column_type="text"),
             ],
         )
-        fake_data_access.set_table_metadata(table_schema)
+        fake_data_access.set_table_metadata_and_return(table_schema)
 
         # table_schema.cql_description =
         user_info = UserInfo(
@@ -438,288 +435,6 @@ class TestChains(unittest.TestCase):
         )
         print(chosen_search_filters)
 
-    # def test_astrapy_selection_chain_summarization(self):
-    #     fake_data_access = DataAccess()
-    #     path_segment_keywords = fake_data_access.get_path_segment_keywords()
-    #     model = ChatOpenAI(model_name="gpt-4-1106-preview")
-    #     user_info_summary = {
-    #         "chain1": "Summary for Customer John Smith (Phone Number: 555-555-5555):\n\n1. Billing Query (26-Dec-2023): John contacted customer support regarding a higher-than-usual bill and was informed of additional charges for international calls he made. The support agent, Sarah, provided a one-time discount as a courtesy, and the issue was resolved.\n\n2. Network Issue (26-Dec-2023): John reported poor network coverage at his home. Mike from technical support checked and found ongoing maintenance work likely causing the issue. Resolution is expected within 48 hours and John will be updated via email. The issue status is pending.\n\n3. Device Support (26-Dec-2023): John sought assistance with setting up his new phone, particularly with software configuration, email setup, and data transfer from his old phone. Laura from device support provided step-by-step guidance, and the issue was resolved successfully.\n\nJohn's interactions with support show he has a history of billing awareness, experiences network-related concerns, and requires assistance with device setup. These details may inform future recommendations for service plans with clearer billing details, network updates, or device setup services.",
-    #         "chain2": "Since you haven't provided any specific information to summarize, I can't create a summary for you. Please provide the details or context that you would like to be summarized, and I'll be happy to help.",
-    #         "chain3": "Summary:\n\nThe Smith family consists of nine members with a range of ages from 16 to 50 years old. Their devices vary from older models like the iPhone 6 to newer ones such as the iPhone 13 and OnePlus 9 Pro. Monthly usage across the family averages between 450 and 800 minutes. The family appears to have multiple support cases, with each member having at least one and some having up to three cases associated with their name. All members share the same primary phone number but have individual numbers for family members. The devices used suggest a mix of iOS and Android preferences within the family.\n\nSpecific member details:\n- John Smith, 29, uses an iPhone 6 and averages 680 minutes monthly.\n- Michael Smith, 42, uses a OnePlus 9 Pro with 800 minutes of usage.\n- Emily Smith, 19, has an iPhone SE and uses around 500 minutes.\n- David Smith, 50, prefers a Google Pixel 6 and uses about 700 minutes a month.\n- Olivia Smith, 24, has a Samsung Galaxy S20, with 600 minutes of usage.\n- William Smith, 33, uses an iPhone 13 and averages 750 minutes.\n- Ava Smith, 16, is on a OnePlus Nord and uses 450 minutes.\n- James Smith, 29, uses a Google Pixel 4a with 680 minutes usage.\n- Sophia Smith, 22, has an iPhone XR and uses approximately 520 minutes.\n\nDevice age and usage patterns could influence recommendations for upgrades or tailored plans to better meet the varying needs of the family members.",
-    #     }
-    #     selected_search_filters = [
-    #         {"metadata.path_segment_1": "support"},
-    #         {"metadata.path_segment_1": "plans"},
-    #         {"metadata.path_segment_3": "international-global-calling"},
-    #         {"metadata.path_segment_3": "account-maintenance-and-management"},
-    #         {
-    #             "metadata.path_segment_3": "device-trade-in-value-how-much-is-my-phone-worth"
-    #         },
-    #         {
-    #             "metadata.path_segment_3": "how-to-use-your-smartphone-as-a-mobile-hotspot"
-    #         },
-    #         {"metadata.path_segment_3": "protect-against-identity-theft"},
-    #         {"metadata.path_segment_3": "smartphone-with-best-camera"},
-    #         {"metadata.path_segment_3": "5g-home-internet-accessories"},
-    #         {"metadata.path_segment_3": "trade-in-value-your-top-questions-answered"},
-    #         {"metadata.path_segment_3": "phone-upgrades-your-top-questions-answered"},
-    #     ]
-    #
-    #     def filtered_ANN_search_helper(_dict):
-    #         return fake_data_access.filtered_ANN_search(
-    #             _dict["collection_filter"], _dict["user_summary"]
-    #         )
-    #
-    #     summarization_chain = (
-    #         RunnableLambda(PromptFactory.clean_string_v2)
-    #         | PromptFactory.build_summarization_prompt(table_schema)
-    #         | model
-    #         | StrOutputParser()
-    #         | RunnableLambda(PromptFactory.clean_string_v2)
-    #         | RunnableLambda(data_access.exec_cql_query_simple)
-    #     )
-    #     # Next, we need to construct the chain for each table.
-    #     chains = [
-    #         ChainFactory.build_summarization_chain(model, fake_data_access, table)
-    #         for table in filtered_tables
-    #     ]
-    #     chain_kwargs = {f"chain{i+1}": chain for i, chain in enumerate(chains)}
-    #     # Pass the dictionary as keyword arguments
-    #     summarization_chains = RunnableParallel(**chain_kwargs)
-    #     summary = summarization_chains.invoke(
-    #         {
-    #             "property_info": user_info,
-    #         }
-    #     )
-    #     search_results = [
-    #         fake_data_access.filtered_ANN_search(search_filter, user_info_summary)
-    #         for search_filter in selected_search_filters
-    #     ]
-    #     print(chosen_search_filters)
-
-    def test_entire_chain(self):
-        fake_data_access = DataAccess()
-        model = ChatOpenAI(model_name="gpt-4-1106-preview")
-        user_info = UserInfo(
-            properties=[
-                PropertyInfo(
-                    property_name="age", property_type="int", property_value=30
-                ),
-                PropertyInfo(
-                    property_name="name",
-                    property_type="text",
-                    property_value="John Smith",
-                ),
-                PropertyInfo(
-                    property_name="phone_number",
-                    property_type="text",
-                    property_value="555-555-5555",
-                ),
-                PropertyInfo(
-                    property_name="email",
-                    property_type="text",
-                    property_value="johndoe@example.com",
-                ),
-                PropertyInfo(
-                    property_name="address",
-                    property_type="text",
-                    property_value="123 Main St, Anytown, USA",
-                ),
-                PropertyInfo(
-                    property_name="account_status",
-                    property_type="text",
-                    property_value="Active",
-                ),
-                PropertyInfo(
-                    property_name="plan_type",
-                    property_type="text",
-                    property_value="Unlimited Data Plan",
-                ),
-            ]
-        )
-
-        factory = ChainFactory()
-        # Pass the dictionary as keyword arguments
-        user_summarization_chain_parallelizable = (
-            factory.build_user_summarization_chain_parallelizable(
-                fake_data_access, model, user_info
-            )
-        )
-        user_info_summary_results = user_summarization_chain_parallelizable.invoke(
-            {"user_info_not_summary": user_info}
-        )
-        # Then, we need to build the chains for
-
-        path_segment_keyword_chain = factory.build_path_segment_keyword_chain(
-            model, user_summarization_chain_parallelizable, fake_data_access
-        )
-
-        filters = path_segment_keyword_chain.invoke(
-            {}
-        )  # [{{"metadata.path_segment_X": "VALUE"}}]
-
-        paired_filters_and_user_summaries = [
-            {"filter": f, "user_info": user_info_summary_results} for f in filters
-        ]
-
-        def run_query(filter_and_summary_pair):
-            collection_filter_inner = filter_and_summary_pair["filter"]
-            user_summary_inner = filter_and_summary_pair["user_info"]
-            search_results_inner = fake_data_access.filtered_ANN_search(
-                collection_filter_inner, user_summary_inner
-            )
-            return search_results_inner
-
-        chains = [
-            (
-                paired_filter_and_user_summary
-                | {"Information": RunnableLambda(run_query)}
-                | PromptFactory.build_summarization_prompt()
-                | model
-                | StrOutputParser()
-            )
-            for paired_filter_and_user_summary in paired_filters_and_user_summaries
-        ]
-
-        chain_kwargs = {f"chain{i+1}": chain for i, chain in enumerate(chains)}
-        # Pass the dictionary as keyword arguments
-        collection_summarization_chain_parallelizable = RunnableParallel(**chain_kwargs)
-
-        ## Final recommendation chain:
-        final_chain = factory.build_final_recommendation_chain(
-            user_summarization_chain_parallelizable,
-            collection_summarization_chain_parallelizable,
-            model,
-        )
-        # Run collection_summarization_chain_parallelizable, user_summarization_chain_parallelizable,
-        # and wrap in a prompt that instructs the user with a recommendation
-
-        message_list = ["Hi, I'm having trouble with my phone network"]
-        final_output = final_chain.invoke(
-            {"user_messages": message_list, "our_responses": [""]}
-        )
-        print(final_output)
-
-    def test_entire_chain_non_parallel(self):
-        fake_data_access = DataAccess()
-        model = ChatOpenAI(model_name="gpt-4-1106-preview")
-        user_info = UserInfo(
-            properties=[
-                PropertyInfo(
-                    property_name="age", property_type="int", property_value=30
-                ),
-                PropertyInfo(
-                    property_name="name",
-                    property_type="text",
-                    property_value="John Smith",
-                ),
-                PropertyInfo(
-                    property_name="phone_number",
-                    property_type="text",
-                    property_value="555-555-5555",
-                ),
-                PropertyInfo(
-                    property_name="email",
-                    property_type="text",
-                    property_value="johndoe@example.com",
-                ),
-                PropertyInfo(
-                    property_name="address",
-                    property_type="text",
-                    property_value="123 Main St, Anytown, USA",
-                ),
-                PropertyInfo(
-                    property_name="account_status",
-                    property_type="text",
-                    property_value="Active",
-                ),
-                PropertyInfo(
-                    property_name="plan_type",
-                    property_type="text",
-                    property_value="Unlimited Data Plan",
-                ),
-            ]
-        )
-        relevant_table_chain = (
-            {
-                "TableList": RunnableLambda(
-                    fake_data_access.get_table_schemas_in_db_v2
-                ),
-                "UserInfo": itemgetter("user_info_not_summary"),
-            }
-            | PromptFactory.build_table_identification_prompt()
-            | model
-            | StrOutputParser()
-            | RunnableLambda(PromptFactory.clean_string_v2)
-            | RunnableLambda(
-                fake_data_access.map_tables_and_populate
-            )  # Not the most performant to rebuild column metadata here, but we can optimize later
-        )
-        relevant_tables = relevant_table_chain.invoke(
-            {"user_info_not_summary": user_info}
-        )
-        factory = ChainFactory()
-        table_summarization_chain_results = []
-        table_summarization_chains = [
-            factory.build_summarization_chain(model, fake_data_access, table)
-            for table in relevant_tables
-        ]
-
-        for chain in table_summarization_chains:
-            result = chain.invoke({"user_info_not_summary": user_info})
-            table_summarization_chain_results.append(result)
-
-        def run_query2(collection_filter_inner, user_summary_inner):
-            def myfunc(empty):
-                print(empty)
-                search_results_inner = fake_data_access.filtered_ANN_search(
-                    collection_filter_inner, user_summary_inner
-                )
-                return search_results_inner
-
-            return myfunc
-
-        collection_summary_chain_result_list = []
-
-        collection_predicate_chain = (
-            factory.build_collection_predicate_chain_non_parallel(
-                model, table_summarization_chain_results, fake_data_access
-            )
-        )
-        collection_predicates = collection_predicate_chain.invoke({})
-        # ^^ [{{"metadata.path_segment_X": "VALUE"}}]
-        for predicate in collection_predicates:
-            vector_search_results = fake_data_access.filtered_ANN_search(
-                predicate, "".join(table_summarization_chain_results)
-            )
-            collection_summary_chain = (
-                {
-                    "Information": RunnableLambda(
-                        run_query2(predicate, table_summarization_chain_results)
-                    ),
-                    "filter": RunnableLambda(lambda x: predicate),
-                    "user_info": RunnableLambda(
-                        lambda x: table_summarization_chain_results
-                    ),
-                }
-                | PromptFactory.build_summarization_prompt()
-                | model
-                | StrOutputParser()
-            )
-            collection_summary_chain_result = collection_summary_chain.invoke({})
-            collection_summary_chain_result_list.append(collection_summary_chain_result)
-            # Are we supposed to compute this summary for every user table? Maybe that's an okay approach.
-        ## Final recommendation chain:
-        final_chain = factory.build_final_recommendation_chain_non_parallel(model)
-        final_output = final_chain.invoke(
-            {
-                "user_summary": table_summarization_chain_results,
-                "business_summary": collection_summary_chain_result_list,
-                "user_messages": ["Hi, I'm having trouble with my phone network"],
-                "our_responses": [""],
-            }
-        )
-
     def test_entire_chain_non_parallel_v2_bak(self):
         fake_data_access = DataAccess()
         model = ChatOpenAI(model_name="gpt-4-1106-preview")
@@ -793,10 +508,8 @@ class TestChains(unittest.TestCase):
             all_user_table_summaries.append(table_summarization)
 
             # Build predicates for table:
-            predicate_idenfication_chain = (
-                factory.build_collection_predicate_chain_non_parallel_v2(
-                    model, table_summarization, all_collection_keywords
-                )
+            predicate_idenfication_chain = factory.build_collection_predicate_chain(
+                model, table_summarization, all_collection_keywords
             )
             collection_predicates = predicate_idenfication_chain.invoke({})
             topic_summaries_for_table = []
@@ -907,7 +620,7 @@ class TestChains(unittest.TestCase):
             all_user_table_summaries.append(table_summarization)
 
             predicate_identification_chain: Runnable = (
-                factory.build_collection_predicate_chain_non_parallel_v2(
+                factory.build_collection_predicate_chain(
                     model, table_summarization, all_collection_keywords
                 )
             )
